+++ src/main/java/com/tenco/team_two_flight_ticket/user/UserController.java
@@ -14,14 +14,12 @@ import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
 import com.tenco.team_two_flight_ticket._core.utils.Define;
+import com.tenco.team_two_flight_ticket.reservation.ReservationResponse.GetMyTravelDTO;
+import com.tenco.team_two_flight_ticket.reservation.ReservationResponse.GetMyTripCountDTO;
 import com.tenco.team_two_flight_ticket.reservation.ReservationResponse.GetMyTripYearDTO;
 import com.tenco.team_two_flight_ticket.reservation.ReservationService;
 import com.tenco.team_two_flight_ticket.user.UserRequest.GetMyTravelListDTO;
-import com.tenco.team_two_flight_ticket.user.UserResponse.GetMyTravelDTO;
 import com.tenco.team_two_flight_ticket.user.UserResponse.GetMyTripCntAndListDTO;
-import com.tenco.team_two_flight_ticket.user.UserResponse.GetMyTripCountDTO;
-
-
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 
@@ -66,9 +64,7 @@ public class UserController {
 
 	@PostMapping("/profile-management")
 	public String updateUser(@Valid UserRequest.UserUpdateDTO dto, Errors errors) {
-
 		userService.updateUser(dto);
-
 		return "redirect:/user/profile";
 	}
 	@GetMapping("/logout")
