+++ src/main/java/com/tenco/team_two_flight_ticket/user/UserResponse.java
@@ -7,6 +7,8 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import com.tenco.team_two_flight_ticket._middle._entity.enums.StatusEnum;
 import com.tenco.team_two_flight_ticket.reservation.ReservationResponse;
+import com.tenco.team_two_flight_ticket.reservation.ReservationResponse.GetMyTravelDTO;
+import com.tenco.team_two_flight_ticket.reservation.ReservationResponse.GetMyTripCountDTO;
 
 import lombok.Data;
 
@@ -21,34 +23,6 @@ public class UserResponse {
         private String scope;
     }
     
-    @Data
-    public static class GetMyTravelDTO {
-    	private int id;
-    	private Boolean isPayed;
-    	private Timestamp createdAt;
-    	private Timestamp paymentDeadline;
-    	private int reservationNum;
-    	private StatusEnum statusEnum;
-    	private String airline;
-    	private String arrivalCity;
-    	private String departureAirport;
-    	private String departureCity;
-    	private String flightName;
-        private Timestamp departureTime;
-        private Timestamp arrivalTime;
-        
-        private String koreanDepartureAirport;
-        private String koreanArrivalAirport;
-        private String koreanAirline;
-        private String koreanDepartureCity;
-    }
-    
-    @Data
-    public static class GetMyTripCountDTO {
-    	private int allTripCount;
-    	private int payedTripCount;
-    	private int notPayedTripCount;
-    }
     
     @Data
     public static class GetMyTripCntAndListDTO {
