+++ src/main/resources/mapper/reservation.xml
@@ -3,7 +3,7 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="com.tenco.team_two_flight_ticket.reservation.ReservationRepository">
-    <select id="getMyTravel" resultType="com.tenco.team_two_flight_ticket.user.UserResponse$GetMyTravelDTO">
+    <select id="getMyTravel" resultType="com.tenco.team_two_flight_ticket.reservation.ReservationResponse$GetMyTravelDTO">
         SELECT r.id as id, is_payed, user_id, r.created_at as created_at, payment_deadline
         , reservation_num, status_enum, airline, arrival_city, departure_airport, arrival_airport , departure_city
         , flight_name, departure_time, arrival_time, icity.name as korean_arrival_city, iairline.name as korean_airline
